pyproject.toml
[project]
name = "greeter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio-tools>=1.71.0",
]
